Add missing docstring
