+++ src/python/array_accessor.h
@@ -4,11 +4,13 @@
 // personal capacity and am not conveying any rights to any intellectual
 // property of any third parties.
 
-#ifndef SRC_PYTHON_ARRAY_ACCESSOR3_H_
-#define SRC_PYTHON_ARRAY_ACCESSOR3_H_
+#ifndef SRC_PYTHON_ARRAY_ACCESSOR_H_
+#define SRC_PYTHON_ARRAY_ACCESSOR_H_
 
 #include <pybind11/pybind11.h>
 
+void addArrayAccessor1(pybind11::module& m);
+void addArrayAccessor2(pybind11::module& m);
 void addArrayAccessor3(pybind11::module& m);
 
-#endif  // SRC_PYTHON_ARRAY_ACCESSOR3_H_
+#endif  // SRC_PYTHON_ARRAY_ACCESSOR_H_
