+++ src/python/cell_centered_vector_grid.cpp
@@ -47,6 +47,7 @@ void addCellCenteredVectorGrid2(py::module& m) {
                  - resolution : Grid resolution.
                  - gridSpacing : Grid spacing.
                  - gridOrigin : Origin point ot the grid.
+                 - domainSizeX : Domain size in x-direction.
              )pbdoc")
         .def_property_readonly(
             "dataSize", &CellCenteredVectorGrid2::dataSize,
@@ -121,6 +122,7 @@ void addCellCenteredVectorGrid3(py::module& m) {
                  - resolution : Grid resolution.
                  - gridSpacing : Grid spacing.
                  - gridOrigin : Origin point ot the grid.
+                 - domainSizeX : Domain size in x-direction.
              )pbdoc")
         .def_property_readonly(
             "dataSize", &CellCenteredVectorGrid3::dataSize,
