+++ src/python/main.cpp
@@ -7,41 +7,77 @@
 #include "advection_solver.h"
 #include "animation.h"
 #include "apic_solver.h"
-#include "array_accessor1.h"
+#include "array_accessor.h"
 #include "bounding_box.h"
+#include "box.h"
 #include "cell_centered_scalar_grid.h"
 #include "cell_centered_vector_grid.h"
 #include "collider.h"
 #include "collocated_vector_grid.h"
+#include "constant_scalar_field.h"
+#include "constant_vector_field.h"
 #include "constants.h"
+#include "cubic_semi_lagrangian.h"
+#include "custom_scalar_field.h"
+#include "custom_vector_field.h"
 #include "cylinder.h"
+#include "eno_level_set_solver.h"
 #include "face_centered_grid.h"
+#include "fdm_cg_solver.h"
+#include "fdm_gauss_seidel_solver.h"
+#include "fdm_iccg_solver.h"
+#include "fdm_jacobi_solver.h"
+#include "fdm_linear_system_solver.h"
 #include "field.h"
 #include "flip_solver.h"
+#include "fmm_level_set_solver.h"
 #include "grid.h"
 #include "grid_backward_euler_diffusion_solver.h"
+#include "grid_blocked_boundary_condition_solver.h"
+#include "grid_boundary_condition_solver.h"
 #include "grid_diffusion_solver.h"
 #include "grid_emitter.h"
 #include "grid_fluid_solver.h"
 #include "grid_forward_euler_diffusion_solver.h"
+#include "grid_fractional_boundary_condition_solver.h"
+#include "grid_fractional_single_phase_pressure_solver.h"
 #include "grid_pressure_solver.h"
+#include "grid_single_phase_pressure_solver.h"
+#include "grid_smoke_solver.h"
 #include "grid_system_data.h"
 #include "implicit_surface.h"
+#include "implicit_triangle_mesh.h"
+#include "iterative_level_set_solver.h"
+#include "level_set_liquid_solver.h"
+#include "level_set_solver.h"
 #include "logging.h"
 #include "particle_emitter.h"
+#include "particle_emitter_set.h"
 #include "particle_system_data.h"
+#include "particle_system_solver.h"
+#include "pci_sph_solver.h"
 #include "physics_animation.h"
 #include "pic_solver.h"
+#include "plane.h"
 #include "point.h"
+#include "point_particle_emitter.h"
 #include "quaternion.h"
 #include "ray.h"
 #include "rigid_body_collider.h"
 #include "scalar_field.h"
 #include "scalar_grid.h"
+#include "semi_lagrangian.h"
 #include "size.h"
+#include "sph_solver.h"
+#include "sph_system_data.h"
 #include "sphere.h"
 #include "surface.h"
+#include "surface_set.h"
+#include "surface_to_implicit.h"
 #include "transform.h"
+#include "triangle.h"
+#include "triangle_mesh.h"
+#include "upwind_level_set_solver.h"
 #include "vector.h"
 #include "vector_field.h"
 #include "vector_grid.h"
@@ -49,44 +85,6 @@
 #include "vertex_centered_vector_grid.h"
 #include "volume_grid_emitter.h"
 #include "volume_particle_emitter.h"
-#include "grid_boundary_condition_solver.h"
-#include "grid_blocked_boundary_condition_solver.h"
-#include "grid_fractional_boundary_condition_solver.h"
-#include "fdm_linear_system_solver.h"
-#include "fdm_jacobi_solver.h"
-#include "fdm_gauss_seidel_solver.h"
-#include "fdm_cg_solver.h"
-#include "fdm_iccg_solver.h"
-#include "grid_single_phase_pressure_solver.h"
-#include "grid_fractional_single_phase_pressure_solver.h"
-#include "grid_smoke_solver.h"
-#include "array_accessor3.h"
-#include "array_accessor2.h"
-#include "box.h"
-#include "triangle.h"
-#include "triangle_mesh.h"
-#include "surface_to_implicit.h"
-#include "surface_set.h"
-#include "plane.h"
-#include "implicit_triangle_mesh.h"
-#include "semi_lagrangian.h"
-#include "cubic_semi_lagrangian.h"
-#include "constant_scalar_field.h"
-#include "custom_scalar_field.h"
-#include "constant_vector_field.h"
-#include "custom_vector_field.h"
-#include "particle_system_solver.h"
-#include "sph_system_data.h"
-#include "sph_solver.h"
-#include "pci_sph_solver.h"
-#include "particle_emitter_set.h"
-#include "point_particle_emitter.h"
-#include "level_set_solver.h"
-#include "iterative_level_set_solver.h"
-#include "eno_level_set_solver.h"
-#include "upwind_level_set_solver.h"
-#include "fmm_level_set_solver.h"
-#include "level_set_liquid_solver.h"
 
 #include <pybind11/pybind11.h>
